requirements.txt
absl-py==0.9.0
addict==2.2.1
aiohttp==3.6.2
alabaster==0.7.12
albumentations==0.4.5
antlr4-python3-runtime==4.8
apex==0.1
appdirs==1.4.3
aspy.yaml==1.3.0
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.3
awscli==1.18.0
Babel==2.9.0
backcall==0.1.0
beautifulsoup4==4.9.0
black==19.3b0
bleach==3.1.0
boto==2.49.0
boto3==1.12.0
botocore==1.15.0
bottle==0.12.18
brainpp==0.6.4.1
bs4==0.0.1
bson==0.5.9
cachetools==4.0.0
certifi==2018.1.18
cffi==1.14.0
cfgv==3.0.0
chardet==3.0.4
cityscapesScripts==2.2.0
Click==7.0
cloudpickle==1.4.1
colorama==0.4.3
coloredlogs==15.0
configparser==5.0.1
confluent-kafka==1.3.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
dataset-format==0.0.0
DCN==1.0
decorator==4.4.1
defusedxml==0.6.0
Deprecated==1.2.7
detectron2==0.3+cu101
dill==0.3.1.1
diskcache==4.1.0
distlib==0.3.0
docker-pycreds==0.4.0
docutils==0.15.2
dulwich==0.19.15
editdistance==0.5.3
einops==0.3.0
elasticsearch==7.5.1
entrypoints==0.3
exp-closure==0.0.0
fairscale==0.1.4
faiss==1.5.3
filelock==3.0.12
fishervector==0.1
flake8==3.8.1
flake8-bugbear==20.11.1
flake8-comprehensions==3.3.1
Flask==1.1.1
flatbuffers==1.12
frozendict==0.4
funcsigs==1.0.2
future==0.15.2
fvcore==0.1.2.post20201104
gevent==1.4.0
gitdb==4.0.5
GitPython==3.1.11
google==2.0.3
google-auth==1.11.2
google-auth-oauthlib==0.4.1
graphviz==0.13.2
greenlet==0.4.15
grpcio==1.27.2
h5py==2.10.0
hp-module==0.0.0
humanfriendly==9.1
hydra-core==1.0.5
icdiff==1.9.1
identify==1.4.11
idna==2.6
idna-ssl==1.1.0
imageio==2.8.0
imagesize==1.2.0
imgaug==0.2.6
importlib-metadata==1.3.0
importlib-resources==1.0.2
influxdb==5.2.3
iotop==0.6
ipdb==0.13.2
ipykernel==5.1.3
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
json5==0.8.5
jsonlines==1.2.0
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
kaggle==1.5.6
kiwisolver==1.1.0
labserver==0.1.0
leven==1.0.4
lmdb==0.98
lxml==4.6.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
MegBrain==8.2.4
MegHair==4.1.1
MegSkull==4.2.0
mistune==0.8.4
mlxtend==0.18.0
mmcv==1.2.5
mmcv-full==1.2.5
mock==4.0.2
more-itertools==8.0.2
msgpack==0.6.2
multidict==4.7.5
name-server==2.2.1
namedlist==1.7
nbconvert==5.6.1
nbformat==5.0.3
neovim==0.2.0
netifaces==0.10.9
networkx==2.4
neupeak==0.3.10
nodeenv==1.3.5
nori==0.9.9
nori2==1.11.1
nose==1.3.7
notebook==6.0.2
numpy==1.17.2
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.1.2.30
packaging==20.1
pandas==0.24.2
pandocfilters==1.4.2
parameterized==0.7.4
parso==0.5.2
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.0.1
pkgconfig==1.5.1
plotly==4.9.0
pluggy==0.13.1
plumbum==1.6.8
portalocker==1.7.0
prat-geventwebsocket==0.9.5
pre-commit==2.0.1
prefetch-generator==1.0.1
pretty-errors==1.2.10
progress==1.5
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.14.0
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycocotools==2.0.2
pycodestyle==2.6.0
pycparser==2.19
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.5.2
pygobject==3.26.1
PyJWT==1.7.1
pykdtree==1.3.1
PyOpenGL==3.1.5
pyparsing==2.4.6
pyquaternion==0.9.9
pyretri==0.1.0+3faf94e
Pyro4==4.79
pyrsistent==0.15.7
PySnooper==0.3.0
pytest==5.3.5
python-dateutil==2.8.1
python-magic==0.4.15
python-slugify==4.0.0
python-statemachine==0.8.0
pytorch-gradcam==0.2.1
pytorch-warmup==0.0.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
randomcolor==0.4.4.5
ray==0.7.0
redis==3.3.11
redispy==3.0.0
requests==2.23.0
requests-futures==0.9.7
requests-oauthlib==1.3.0
resnest==0.0.5
retrying==1.3.3
rrun==1.7.7
rsa==3.4.2
s3transfer==0.3.3
scikit-image==0.14.3
scikit-learn==0.22.1
scipy==1.2.1
seaborn==0.11.0
Send2Trash==1.5.0
sentry-sdk==0.19.4
serpent==1.30.2
setproctitle==1.1.10
Shapely==1.6.4.post2
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smart-open==1.9.0
smmap==3.0.4
snowballstemmer==2.1.0
sortedcontainers==2.1.0
soupsieve==2.0
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
ssh-import-id==5.7
subprocess32==3.5.4
tabulate==0.8.6
tensorboard==1.15.0
tensorboard-logger==0.1.0
tensorboardX==2.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
text-unidecode==1.3
texttable==1.6.2
token-bucket==0.2.0
toml==0.10.0
torch==1.6.0+cu101
torchfile==0.1.0
torchlars==0.1.2
torchnet==0.0.5.1
torchsummary==1.5.1
torchvision==0.7.0+cu101
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
ttach==0.0.2
tushare==1.2.62
typeguard==2.7.1
typing==3.7.4.1
typing-extensions==3.7.4.2
ujson==1.35
urllib3==1.22
virtualenv==20.0.4
visdom==0.1.8.9
waitress==1.4.2
Wand==0.5.9
wandb==0.10.11
watchdog==0.10.4
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.12.0
yacs==0.1.6
yapf==0.30.0
yarl==1.4.2
zipp==0.6.0
